requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
nudenet>=3.4.0
opencv-python-headless>=4.8.0
aiohttp>=3.8.0
aiofiles>=23.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiosqlite>=0.19.0
Pillow>=10.0.0
onnxruntime>=1.16.0
faiss-cpu>=1.7.0
